refactor stats and ci
