+++ src/header.ts
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         OpenAI Codex Enhancer
 // @namespace    http://tampermonkey.net/
-// @version      1.0.54
+// @version      1.0.55
 // @description  Enhances user experience inside OpenAI Codex
 // @match        https://chatgpt.com/codex*
 // @grant        GM_xmlhttpRequest
